# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dccba68f1ef4518a53834d0
  title: wishbox
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /post/:
    post:
      x-skaffolder-id: 5dccbdc3f1ef4518a53835a8
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dccbd47f1ef4518a5383576
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Post
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                createdAt:
                  type: integer
                title:
                  type: string
                updatedAt:
                  type: integer
    get:
      x-skaffolder-id: 5dccbdc3f1ef4518a53835ab
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dccbd47f1ef4518a5383576
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Post
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /post/{id}:
    delete:
      x-skaffolder-id: 5dccbdc3f1ef4518a53835ac
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dccbd47f1ef4518a5383576
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Post
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dccbdc3f1ef4518a53835a5
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dccbd47f1ef4518a5383576
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Post
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    post:
      x-skaffolder-id: 5dccbdc3f1ef4518a53835a7
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dccbd47f1ef4518a5383576
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Post
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post"
  /post/findBycreatedBy/{key}:
    get:
      x-skaffolder-id: 5dccbdb3f1ef4518a538357d
      x-skaffolder-name: findBycreatedBy
      x-skaffolder-id-resource: 5dccbd47f1ef4518a5383576
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: findBycreatedBy
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBycreatedBy
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBycreatedBy/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBycreatedBy
      tags:
        - Post
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa createdBy da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    
  #-----  Tag  -----
  /tag/:
    post:
      x-skaffolder-id: 5dccbdb7f1ef4518a5383586
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dccbbedf1ef4518a538355b
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Tag
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
    get:
      x-skaffolder-id: 5dccbdb7f1ef4518a538358b
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dccbbedf1ef4518a538355b
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Tag
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
  /tag/{id}:
    delete:
      x-skaffolder-id: 5dccbdb7f1ef4518a538358c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dccbbedf1ef4518a538355b
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dccbdb7f1ef4518a5383585
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dccbbedf1ef4518a538355b
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    post:
      x-skaffolder-id: 5dccbdb7f1ef4518a5383583
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dccbbedf1ef4518a538355b
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tag"
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Post:
      x-skaffolder-id: 5dccbd47f1ef4518a5383576
      x-skaffolder-id-db: 5dccba68f1ef4518a53834d1
      x-skaffolder-id-entity: 5dccbd47f1ef4518a5383570
      x-skaffolder-url: /post
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        content:
          type: string
          x-skaffolder-id-attr: 5dccbd47f1ef4518a5383572
          x-skaffolder-type: String
          x-skaffolder-required: true
        createdAt:
          type: integer
          x-skaffolder-id-attr: 5dccbd47f1ef4518a5383573
          x-skaffolder-type: Date
        title:
          type: string
          x-skaffolder-id-attr: 5dccbd47f1ef4518a5383571
          x-skaffolder-type: String
          x-skaffolder-required: true
        updatedAt:
          type: integer
          x-skaffolder-id-attr: 5dccbd47f1ef4518a5383574
          x-skaffolder-type: Date
      x-skaffolder-relations: 
        createdBy:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dccbd47f1ef4518a5383570
          x-skaffolder-ent2: 5dccba68f1ef4518a53834d2
      required:
        - _id
        - content
        - title
    Tag:
      x-skaffolder-id: 5dccbbedf1ef4518a538355b
      x-skaffolder-id-db: 5dccba68f1ef4518a53834d1
      x-skaffolder-id-entity: 5dccbbedf1ef4518a5383558
      x-skaffolder-url: /tag
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5dccbbedf1ef4518a5383559
          x-skaffolder-type: String
          x-skaffolder-required: true
        type:
          type: string
          x-skaffolder-id-attr: 5dccbbedf1ef4518a538355a
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
        - name
...